Fix bugs in 'sticky' ordering of news posts
Don't attempt to order by meta_value with an EXISTS / NOT EXISTS clause,
because the resulting SQL is wack and may end up trying to order by some
other arbitrary meta value.
Use a True/False key for the sticky status, instead of a checkbox one,
because checkbox fields store their values in serialized arrays, and
ordering by those is a bad idea.